Confirm that createWallet works
