+++ app/actions/ClientActions.js
@@ -1,4 +1,4 @@
-import { getBalance, getStakeInfo } from '../middleware/grpc/client';
+import { getBalance, createWallet } from '../middleware/grpc/client';
 
 export const GETBALANCE_ATTEMPT = 'GETBALANCE_ATTEMPT';
 export const GETBALANCE_FAILED = 'GETBALANCE_FAILED';
@@ -29,6 +29,43 @@ export function grpcBalance() {
         dispatch(getBalanceError(err + " Please try again"));
       } else {
         dispatch(getBalanceSuccess(balance));
+      }
+    })
+  }
+}
+
+export const CREATEWALLET_ATTEMPT = 'CREATEWALLET_ATTEMPT';
+export const CREATEWALLET_FAILED = 'CREATEWALLET_FAILED';
+export const CREATEWALLET_SUCCESS = 'CREATEWALLET_SUCCESS';
+
+function createWalletError(error) {
+  return { error, type: CREATEWALLET_FAILED };
+}
+
+function createWalletSuccess() {
+  return { type: CREATEWALLET_SUCCESS };
+}
+
+export function createWalletRequest(pubPass, privPass, seed) {
+  console.log("blah", pubPass, privPass, seed);
+  return { 
+      pubPass: pubPass,
+      privPass: privPass,
+      seed: seed,
+      type: CREATEWALLET_ATTEMPT };
+}
+
+export function createNewWallet() {
+  return (dispatch, getState) => {
+    const { client } = getState().login;
+    const { pubPass, privPass, seed } = getState().grpc;
+    console.log(pubPass, privPass, seed);
+    createWallet(client, pubPass, privPass, seed, 
+        function(err) {
+      if (err) {
+        dispatch(createWalletError(err + " Please try again"));
+      } else {
+        dispatch(createWalletSuccess());
       }
     })
   }
